Rename UpdateOrderStatusResult to UpdateOrderResult
